Add group selector for button classes
Both buttons in index.html use the classes 'button', 'confirm', and
'cancel'. As they share similar characteristics (e.g. font-size,
dimensions), they were put in a group selector.

Outstanding issue: The lone 'confirm' class is supposed to take
precedence over the group selector but is currently not. This will be
handled in a future commit.